get rid of int cast deprecation warning
2015-04-18 19:10:51,515 WARNING hex:145 /usr/local/lib/python3.4/dist-packages/qtile-0.9.1-py3.4.egg/libqtile/utils.py:145: DeprecationWarning: automatic int conversions have been deprecated
  return '#%02x%02x%02x' % (r * 255, g * 255, b * 255)

+++ libqtile/utils.py
@@ -142,7 +142,7 @@ def rgb(x):
 
 def hex(x):
     r, g, b, _ = rgb(x)
-    return '#%02x%02x%02x' % (r * 255, g * 255, b * 255)
+    return '#%02x%02x%02x' % (int(r * 255), int(g * 255), int(b * 255))
 
 
 def scrub_to_utf8(text):
